[debian-jessie-small] Use policy-rc.d to block service invocation.
Temporarily add a /usr/sbin/policy-rc.d file returning exit code 101 to
disallow service invocation during the build stage. While not strictly
necessary given the short list of packages installed by this profile, it
is useful to have it in place in case the package list is expanded
later, or when using this profile as a basis for others.

See /usr/share/doc/sysv-rc/README.policy-rc.d.gz for details about the
policy-rc.d file itself.