[debian-jessie-split] Copy executables to initrd.
Copy executables and related libraries from the root filesystem to the
initrd image. The mechanism includes package download, use of the
reslibs script, and careful inclusion of libnss_dns to support DNS
resolution through libc.

This approach is very prone to portability problems.

The 'strace' executable is included, as it is invaluable for debugging
init problems.