[debian-jessie-zram] Add zram overlay option.
Add zram support for the read-write overlay filesystem. This uses the
zram kernel module and initiates four zram devices by default. The first
one (zram0) is configured with a size of twice the system's memory size,
formatted as ext4 without a journal, and then mounted with options
'discard' (to free up memory when deleting files) and 'barrier=0' (which
is only relevant for journalling).

Add support for the 'rib.rdtype' kernel parameter to select between
'tmpfs' (the default), and 'zram'.